# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Quoter API
  version: 0.1.0
  description: Self-hosted Go/Gin API for the Quoter iPad app.
servers:
  - url: http://localhost:8080/api/v1
paths:
  /auth/register:
    post:
      summary: Register a company owner account.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                company_name: { type: string }
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 10 }
      responses:
        "200": { description: Auth tokens and user profile }
  /auth/login:
    post:
      summary: Login with email and password.
      responses:
        "200": { description: Auth tokens and user profile }
  /auth/refresh:
    post:
      summary: Rotate a refresh token and issue a new access token.
      responses:
        "200": { description: New token pair }
  /auth/logout:
    post:
      security: [{ bearerAuth: [] }]
      summary: Revoke a refresh token.
      responses:
        "204": { description: Logged out }
  /auth/me:
    get:
      security: [{ bearerAuth: [] }]
      summary: Return the current user and company.
      responses:
        "200": { description: Current session }
  /estimate-templates:
    get:
      security: [{ bearerAuth: [] }]
      summary: List saved contractor estimate templates.
      responses:
        "200":
          description: Estimate template list.
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/EstimateTemplate"
    post:
      security: [{ bearerAuth: [] }]
      summary: Save or replace a named contractor estimate template.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateTemplateInput"
      responses:
        "200": { description: Existing template replaced }
        "201": { description: New template created }
  /estimate-templates/{id}:
    get:
      security: [{ bearerAuth: [] }]
      summary: Get a saved contractor estimate template.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Estimate template }
    put:
      security: [{ bearerAuth: [] }]
      summary: Update a saved contractor estimate template.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EstimateTemplateInput"
      responses:
        "200": { description: Estimate template updated }
    delete:
      security: [{ bearerAuth: [] }]
      summary: Delete a saved contractor estimate template.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: Estimate template deleted }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    EstimateTemplate:
      type: object
      properties:
        id: { type: string, format: uuid }
        project_id: { type: string, format: uuid }
        name: { type: string }
        renovation_type: { type: string }
        categories:
          type: array
          items:
            type: object
            additionalProperties: true
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    EstimateTemplateInput:
      type: object
      required: [name, renovation_type, categories]
      properties:
        project_id: { type: string, format: uuid }
        name: { type: string }
        renovation_type: { type: string }
        categories:
          type: array
          items:
            type: object
            additionalProperties: true